Rename track member variable. Tidy up test file and specify name.
